+++ tests/hash-graph-integration/postgres/links.rs
@@ -1,5 +1,5 @@
 use graph_test_data::{data_type, entity, entity_type, property_type};
-use graph_types::knowledge::{entity::EntityProperties, link::EntityLinkOrder};
+use graph_types::knowledge::entity::EntityProperties;
 use type_system::url::{BaseUrl, OntologyTypeVersion, VersionedUrl};
 
 use crate::DatabaseTestWrapper;
@@ -254,17 +254,9 @@ async fn remove_link() {
             .is_empty()
     );
 
-    api.archive_entity(
-        link_entity_metadata.record_id.entity_id,
-        EntityProperties::empty(),
-        vec![friend_link_type_id],
-        EntityLinkOrder {
-            left_to_right: None,
-            right_to_left: None,
-        },
-    )
-    .await
-    .expect("could not remove link");
+    api.archive_entity(link_entity_metadata.record_id.entity_id)
+        .await
+        .expect("could not remove link");
 
     assert!(
         api.get_latest_entity_links(alice_metadata.record_id.entity_id)