+++ libs/@hashintel/query-editor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hashintel/query-editor
 
+## 0.0.3
+
+### Patch Changes
+
+- [#2966](https://github.com/hashintel/hash/pull/2966) [`4c30e84ea`](https://github.com/hashintel/hash/commit/4c30e84ea521812e9ea41299afbe02c33787a886) Thanks [@CiaranMn](https://github.com/CiaranMn)! - update react-hook-form
+
+- Updated dependencies [[`34e770e21`](https://github.com/hashintel/hash/commit/34e770e2163eaf7a1ed60818d148cc17dba83d93), [`4c30e84ea`](https://github.com/hashintel/hash/commit/4c30e84ea521812e9ea41299afbe02c33787a886), [`964f46f70`](https://github.com/hashintel/hash/commit/964f46f709198c836b6e83f478ef0962a3de1d7e), [`5a45182d1`](https://github.com/hashintel/hash/commit/5a45182d18e69681b8f69f6220fdcd8573bfaf7e), [`890d138d8`](https://github.com/hashintel/hash/commit/890d138d8523a7c9604786094a4156e19e638f09), [`765cc9816`](https://github.com/hashintel/hash/commit/765cc98162d5208eca9a4c6dd753b3ff28a4d0f8), [`2594988cc`](https://github.com/hashintel/hash/commit/2594988cc6078318847f5ec96d19d03f5c082b04)]:
+  - @hashintel/design-system@0.0.8
+
 ## 0.0.2
 
 ### Patch Changes